Renames files after code refactor
