+++ src/eatmemory.apple.c
@@ -1,4 +1,4 @@
-#include <sysmem.h>
+#include <eatmemory.h>
 #ifdef SYSMEM_MODE_APPLE
     #include <mach/mach.h>
     #include <mach/mach_host.h>
