+++ src/eatmemory.linux.c
@@ -1,4 +1,4 @@
-#include <sysmem.h>
+#include <eatmemory.h>
 #ifdef SYSMEM_MODE_LINUX
 #include <stdio.h>
 #include <unistd.h>
